# RFC: allow splatting in hvcat syntax