This changes the lowering of hvcat syntax to properly support splatting, which also simplified lowering slightly. I was a bit worried about introducing another function call here in terms of type inference/constant prop, but in my initial micro benchmarks, I don't see any regression:
master:
```julia
julia> @time [1 2 3 4; 5 6 7 8; 9 10 11 12; 13 14 15 16]
  0.075053 seconds (36.30 k allocations: 2.334 MiB, 99.99% compilation time)
4×4 Matrix{Int64}:
  1   2   3   4
  5   6   7   8
  9  10  11  12
 13  14  15  16

julia> @time [1 2 3 4; 5 6 7 8; 9 10 11 12; 13 14 15 16]
  0.000008 seconds (2 allocations: 352 bytes)
4×4 Matrix{Int64}:
  1   2   3   4
  5   6   7   8
  9  10  11  12
 13  14  15  16

julia> @time [[1, 2] [3 4; 5 6]; [7 8] 9]
  0.966912 seconds (964.36 k allocations: 57.004 MiB, 4.15% gc time, 99.93% compilation time)
3×3 Matrix{Int64}:
 1  3  4
 2  5  6
 7  8  9

julia> @time [[1, 2] [3 4; 5 6]; [7 8] 9]
  0.000108 seconds (49 allocations: 2.453 KiB)
3×3 Matrix{Int64}:
 1  3  4
 2  5  6
 7  8  9
```
This PR:
```julia
julia> @time [1 2 3 4; 5 6 7 8; 9 10 11 12; 13 14 15 16]
  0.064680 seconds (36.30 k allocations: 2.334 MiB, 99.99% compilation time)
4×4 Matrix{Int64}:
  1   2   3   4
  5   6   7   8
  9  10  11  12
 13  14  15  16

julia> @time [1 2 3 4; 5 6 7 8; 9 10 11 12; 13 14 15 16]
  0.000010 seconds (2 allocations: 352 bytes)
4×4 Matrix{Int64}:
  1   2   3   4
  5   6   7   8
  9  10  11  12
 13  14  15  16

julia> @time [[1, 2] [3 4; 5 6]; [7 8] 9]
  0.847476 seconds (963.75 k allocations: 56.888 MiB, 3.57% gc time, 99.94% compilation time)
3×3 Matrix{Int64}:
 1  3  4
 2  5  6
 7  8  9

julia> @time [[1, 2] [3 4; 5 6]; [7 8] 9]
  0.000080 seconds (49 allocations: 2.453 KiB)
3×3 Matrix{Int64}:
 1  3  4
 2  5  6
 7  8  9
```
fixes #38844