@JeffBezanson 2021-01-14
I wonder if it would be better just to wrap `t...` as `permutedims([t...])`. I think that would work? Might help avoid pathological tuple cases.