@simeonschaub 2021-01-15
Do you mean during lowering? That would always be allocating though and wouldn' t work for tuples of arrays, for example. Could you clarify what you mean by the pathological tuple cases here?