@JeffBezanson 2021-01-15
Oh right it would actually just be `hcat(t...)`. The worry is that tuple_cat is O(n^2), and that this approach will generate lots of new tuple types.