@simeonschaub 2021-01-15
Would something like a fallback method for `Any16` for `tuple_cat` that avoids this overspecialization by building up an intermediate `Vector{Any}` instead address your worries? How worried are we actually about scaling here? Are people really using this for constructing huge arrays?