@simeonschaub 2021-01-25
I have no idea what is going on here:
```scheme
(expand-forms '(vcat (row a (|...| b)) (row c d)))
(call (top hvcat_rows) (call (core _apply_iterate) (top iterate) (core tuple)
			    (call (core tuple) a) b)
      (call (core tuple) c d))

(expand-forms '(vcat (row a (|...| b)) (row c d)))
(call (top hvcat_rows) (call (core _apply_iterate) (top iterate) (core tuple)
			    (call (core tuple) a) b)
      (call (core tuple) c d) (call (core _apply_iterate) (top iterate) (core
  tuple)
				   (call (core tuple) a) b)
      (call (core tuple) c d))

(expand-forms '(vcat (row a (|...| b)) (row c d)))
(call (top hvcat_rows) (call (core _apply_iterate) (top iterate) (core tuple)
			    (call (core tuple) a) b)
      (call (core tuple) c d) (call (core _apply_iterate) (top iterate) (core
  tuple)
				   (call (core tuple) a) b)
      (call (core tuple) c d) (call (core _apply_iterate) (top iterate) (core
  tuple)
				   (call (core tuple) a) b)
      (call (core tuple) c d))
```
It's probably just a dumb mistake I made, but I am running out of ideas. Weirdly enough, it works just fine with `typed_vcat`. @JeffBezanson Any idea what is going on here?