@BioTurboNick 2021-01-25
Just saw you're working on this! Cool, looking at it myself.

Based on the checks, it looks like your PR is not compatible with this line from base\uuid.jl, line 84:
`[36:-1:25; 23:-1:20; 18:-1:15; 13:-1:10; 8:-1:1]`, so I can't build your PR right now.