@simeonschaub 2021-01-25
Yes, that's the problem I described above. I am increasingly suspecting this might actually be a femtolisp bug. Perhaps something similar to https://github.com/JeffBezanson/femtolisp/issues/43?