@BioTurboNick 2021-01-25
Ah! I wasn't clear on what you were showing. Is it that repeated execution of the same line is causing different output?