@simeonschaub 2021-01-25
Exactly. This is taken from Julia's flisp repl