@BioTurboNick 2021-01-25
I think found a parser solution that doesn't require new functions.

```
a = [1,2,3]
b = [4,5,6]
[a... b...; b... a...]
```

It's effectively doing this: `hvcat((sum((length(a),length(b))), sum((length(a), length(b)))), a..., b..., b..., a...)`

`@btime` gives 702.098 ns (7 allocations: 448 bytes), doesn't have any impact on the standard hvcats (4 allocations: 288 bytes).

Wondering if there's a way to get rid of the extra allocations. How does it compare to yours?

```
   'vcat
   (lambda (e)
     (let ((a (cdr e)))
       (if (any assignment? a)
           (error (string "misplaced assignment statement in \"" (deparse e) "\"")))
       (if (has-parameters? a)
           (error "unexpected semicolon in array expression"))
       (expand-forms
         (if (any (lambda (x)
                   (and (pair? x) (eq? (car x) 'row)))
                   a)
           ;; convert nested hcat inside vcat to hvcat
           (let ((rows (map (lambda (x)
                              (if (and (pair? x) (eq? (car x) 'row))
                                (cdr x)
                                (list x)))
                              a)))
             (let ((lengths (map (lambda (r)
                                   `(call (top sum) (tuple
                                     ,.(map (lambda (x)
                                            (if (vararg? x)
                                              `(call (top length) ,.(cdr x))
                                              1))
                                          r))))
                                 rows)))
               `(call (top hvcat)
                      (tuple ,.lengths)
                      ,.(apply append rows))))
             `(call (top vcat) ,@a)))))
```