@simeonschaub 2021-01-25
The problem with that is that not all iterators define `length`. Since splatting should work with any iterator, I don't think we can do this.