@BioTurboNick 2021-01-25
Though you'd need iterators to have matching numbers of elements for an hvcat operation to work at all... Is there an example of where this would be an issue?