@simeonschaub 2021-01-25
I will admit, this is a little contrived for `hvcat`, but I am sure there are other examples as well:
```
julia> write("foo", "foo\nbar")
7

julia> itr = eachline("foo")
Base.EachLine{IOStream}(IOStream(<file foo>), Base.var"#323#324"{IOStream}(IOStream(<file foo>)), false)

julia> string(itr...)
"foobar"

julia> length(itr)
ERROR: MethodError: no method matching length(::Base.EachLine{IOStream})
Closest candidates are:
  length(::BitSet) at bitset.jl:365
  length(::LibGit2.GitStatus) at /buildworker/worker/package_linuxaarch64/build/usr/share/julia/stdlib/v1.5/LibGit2/src/status.jl:21
  length(::Base.Iterators.Flatten{Tuple{}}) at iterators.jl:1061
  ...
Stacktrace:
 [1] top-level scope at REPL[17]:1
```