@BioTurboNick 2021-01-26
So I don't know _why_ it's a problem, but Julia/flisp doesn't like that `(top vcat)` etc. are being assigned to variables like that. If I place them explicitly inside the `call`, it builds fine.

(Also I don't know what this does, but I think the last line should have `,@a` instead of `,.a` ? At least that's what it is in the original.)