@BioTurboNick 2021-01-26
I basically adapted your solution into FLISP, looks like it is a bit more efficient. Let me know what you think.

```
a = [1,2,3]
b = [4,5,6]
@btime [a... b...; b... a...]
#=
  848.611 ns (6 allocations: 480 bytes)
2×6 Matrix{Int64}:
 1  2  3  4  5  6
 4  5  6  1  2  3
=#

@btime Base.hvcat_rows(tuple(a..., b...), tuple(b..., a...))
#=
  1.043 μs (8 allocations: 704 bytes)
2×6 Matrix{Int64}:
 1  2  3  4  5  6
 4  5  6  1  2  3
=#
```

One reason I'm interested in a solution that doesn't require extra methods is so I can adapt it to my N-dimensional syntax PR. Checked it with your `eachline` example to ensure it worked on length-less iterators.

```
(let ((a (cdr e)))
    (if (any assignment? a)
        (error (string "misplaced assignment statement in \"" (deparse e) "\"")))
    (if (has-parameters? a)
        (error "unexpected semicolon in array expression"))
    (expand-forms
      (if (any (lambda (x)
                (and (pair? x) (eq? (car x) 'row)))
                a)
        ;; convert nested hcat inside vcat to hvcat
        (let ((rows (map (lambda (x)
                          (if (and (pair? x) (eq? (car x) 'row))
                            (cdr x)
                            (list x)))
                          a)))
          ;; in case there is splatting inside `hvcat`, collect each row as a tuple for length determination
          (let ((has-vararg (any (lambda (row) (any vararg? row)) rows)))
            (let ((lengths (cond (has-vararg (map (lambda (x) `(call (top length) ,x))
                                                  (map (lambda (x) `(tuple ,.x)) rows)))
                                 (else       (map length rows)))))
              `(call (top hvcat)
                     (tuple ,.lengths)
                     ,.(apply append rows)))))
        `(call (top vcat) ,@a)))))
```