@simeonschaub 2021-01-26
Hmm, I think the problem with that approach is that you are iterating the splatted iterator twice, which is a problem if the iterator is stateful. I am actually surprised this works for `eachline`, does it actually return the correct result?